sudokusolver.c
As stated in the main description, Sudoku Solver works by testing numbers to see if they are valid moves or not. If they are valid, it moves onto another test until either it has reached a contradiction or the puzzle has found a possible solution. It only finds one possible solution if there are multiple.